RiemannNoise

+++ net.py
@@ -735,9 +735,9 @@ class OptimizedBlock(nn.Module):
         self.downsample = nn.AvgPool2d(2) if downsample else nn.Identity()
 
     def init_spectral_norm(self):
-        self.conv_1 = spectral_norm(self.conv_1).to(device,dtype=torch.half)
-        self.conv_2 = spectral_norm(self.conv_2).to(device,dtype=torch.half)
-        self.c_sc = spectral_norm(self.c_sc).to(device,dtype=torch.half)
+        self.conv_1 = spectral_norm(self.conv_1).to(device)
+        self.conv_2 = spectral_norm(self.conv_2).to(device)
+        self.c_sc = spectral_norm(self.c_sc).to(device)
 
     def forward(self, in_feat):
         x = self.conv_1(in_feat)
@@ -835,9 +835,9 @@ def calc_GAN_loss_from_pred(prediction: torch.Tensor,
     c = 64
     h = 256
     if target_is_real:
-        target_tensor = torch.ones(batch_size, c, h, h, device=torch.device('cuda:0'), dtype=torch.float)
+        target_tensor = torch.ones(batch_size, c, h, h, device=torch.device('cuda:0'))
     else:
-        target_tensor = torch.zeros(batch_size, c, h, h,device=torch.device('cuda:0'), dtype=torch.float)
+        target_tensor = torch.zeros(batch_size, c, h, h,device=torch.device('cuda:0'))
     loss = F.mse_loss(prediction, target_tensor.detach())
     return loss
